# openapi/v1.yaml
components:
  responses:
...
    Unauthorized:
      content: *id001
      description: Missing, invalid, expired or revoked credential.
      headers:
        WWW-Authenticate:
          schema:
            type: string
  schemas:
...
    ApiKeyOut:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        expires_at:
          format: date-time
          title: Expires At
          type: string
        id:
          format: uuid
          title: Id
          type: string
        name:
          title: Name
          type: string
        owner_id:
          title: Owner Id
          type: string
        public_id:
          title: Public Id
          type: string
        revoked_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Revoked At
        roles:
          items:
            type: string
          title: Roles
          type: array
        scopes:
          items:
            type: string
          title: Scopes
          type: array
      required:
      - id
      - public_id
      - name
      - owner_id
      - roles
      - scopes
      - created_at
      - expires_at
      - revoked_at
      title: ApiKeyOut
      type: object
...
  securitySchemes:
    apiKey:
      bearerFormat: nfb_live_<public id>_<secret>
      description: Scoped, expiring API key (SEC-014). Revocation is effective at once for new requests
        and within 60 s for open streams (SEC-017).
      scheme: bearer
      type: http
    oidc:
      description: 'User access token from the identity provider (Authorization: Bearer).'
      openIdConnectUrl: https://id.example.invalid/.well-known/openid-configuration
      type: openIdConnect
info:
  description: 'Platform API v1. Data flows device -> SDK -> platform only: no operation, event or field
    sends anything to acquisition hardware (SEC-090). Errors are RFC 9457 application/problem+json. Only
    additive changes within v1; deprecations carry Deprecation and Sunset headers.'
  title: Platform API
  version: 1.0.0
openapi: 3.1.0
paths:
  /v1/api-keys:
    get:
      operationId: listApiKeys
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiKeyOut'
                title: Response Listapikeys
                type: array
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: List Api Keys
      x-nf-action: apikey:read
    post:
      operationId: createApiKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyIssued'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: Create Api Key
      x-nf-action: apikey:create
  /v1/api-keys/{key_id}:
    delete:
      operationId: revokeApiKey
      parameters:
      - in: path
        name: key_id
        required: true
        schema:
          format: uuid
          title: Key Id
          type: string
      responses:
        '204':
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: Revoke Api Key
      x-nf-action: apikey:revoke
...
  /v1/audit/batches:
    get:
      operationId: listAuditBatches
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditBatchOut'
                title: Response Listauditbatches
                type: array
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: List Audit Batches
      x-nf-action: audit:read
  /v1/audit/batches/{seq}/object:
    get:
      description: 'AppSec M1: the exact bytes of one of the caller''s tenant''s audit batches, as stored
        in the

        WORM ``audit`` bucket, so an auditor can recompute the batch ID and the chain links

        independently (``X-NF-Batch-Id`` is the ID the platform recorded).'
      operationId: getAuditBatchObject
      parameters:
      - in: path
        name: seq
        required: true
        schema:
          minimum: 0
          title: Seq
          type: integer
      responses:
        '200':
          content:
            application/json: {}
          description: the stored batch bytes (NF-CJSON; the ID is SHA-256 over them, hashing spec v2
            section 9.1)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: Get Audit Batch Object
      x-nf-action: audit:read
  /v1/audit/events:
    get:
      operationId: listAuditEvents
      parameters:
      - in: query
        name: after_seq
        required: false
        schema:
          default: 0
          minimum: 0
          title: After Seq
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditEventOut'
                title: Response Listauditevents
                type: array
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: List Audit Events
      x-nf-action: audit:read
...
  /v1/datasets/{dataset_id}:
    get:
      operationId: getDataset
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          format: uuid
          title: Dataset Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Get Dataset
      x-nf-action: dataset:read
      x-nf-api-key-scope: metadata:read
  /v1/datasets/{dataset_id}/subjects:
    get:
      operationId: listSubjects
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          format: uuid
          title: Dataset Id
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SubjectOut'
                title: Response Listsubjects
                type: array
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: List Subjects
      x-nf-action: subject:read
      x-nf-api-key-scope: metadata:read
    post:
      operationId: createSubject
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          format: uuid
          title: Dataset Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Create Subject
      x-nf-action: subject:create
      x-nf-api-key-scope: metadata:write
...
  /v1/whoami:
    get:
      operationId: whoami
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhoAmIOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Whoami
      x-nf-action: self:read
servers:
- url: https://api.example.invalid
...